+++ activerecord/lib/active_record/fixture_set/file.rb
@@ -18,23 +18,34 @@ module ActiveRecord
       def initialize(file)
         @file = file
         @rows = nil
+        @raw_rows = nil
+        @model_class = nil
       end
 
       def each(&block)
         rows.each(&block)
       end
 
+      def model_class
+        return @model_class if @model_class
+        row = raw_rows.find { |fixture_name, _| fixture_name == '_fixture' }
+        @model_class = row.last['model_class'] if row
+      end
 
       private
         def rows
-          return @rows if @rows
+          @rows ||= raw_rows.reject { |fixture_name, _| fixture_name == '_fixture' }
+        end
+
+        def raw_rows
+          return @raw_rows if @raw_rows
 
           begin
             data = YAML.load(render(IO.read(@file)))
           rescue ArgumentError, Psych::SyntaxError => error
             raise Fixture::FormatError, "a YAML error occurred parsing #{@file}. Please note that YAML must be consistently indented using spaces. Tabs are not allowed. Please have a look at http://www.yaml.org/faq.html\nThe exact error was:\n  #{error.class}: #{error}", error.backtrace
           end
-          @rows = data ? validate(data).to_a : []
+          @raw_rows = data ? validate(data).to_a : []
         end
 
         def render(content)